Increased the chance of Katya's first mission spawning
I am currently two years into a playthrough, am running the second Tarazed convoy...and Katya hasn't held a gun to my head yet. So I went to take a look at the chance of that mission actually spawning. 

Given how much of the game content is locked behind her doing so, and how weird the storyline gets the longer you take about it, and how it seems to presume you will do the quest in the early game, there is zero reason for its chance of spawning to be so low. Hell, the only reason I wouldn't want to set it to 100 is because I think there's some merit to it being even more common if the Free Worlds likes you. 

This assumes I have not misunderstood the chance of it spawning, and that higher numbers mean a higher chance and lower numbers mean a lower one.